requirements.txt
tensorflow>=2.10.0
flask>=2.0.0
numpy>=1.21.0
Pillow>=9.0.0
h5py>=3.6.0
werkzeug>=2.0.0
scipy>=1.7.0